requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
pytest-playwright
jinja2